Update
This brings an improved version of the previous CreateTI script with more options to create the output. It's currently being applied to the rerun of G4 structures. It still needs to be updated, but it's closer to the initial thing I wanted to make but I wasn't able to work on.