Update
This major functionality update adds various new functionality. Most recently, today, I created the newest LoadFiles() function which is supposed to create less overhead when creating workspace variables. The next part of this update is to erase all parts that recreate the  K_Means algorithm and replace them with the saved indices in the global variables KMEANS.