Small Update
Added descriptions to several of the functions and fixed the issue where global variables weren't being loaded to functions (it was caused due to the clear line at the beginning of script).

I was able to remove the lines where the K-means clustering was being executed since those results can be saved beforehand.

More descriptions need to be fully finished.