Update: GetBonds
GetBonds.tcl: Created a new TCL script to extract the BondLengths of files consisting of Guanine Pairs. Because of this structure, the output function, WriteCSV, writes in a very specific format (i.e. It uses 2 loops to write a line onto the CSV).  The main loop in GetGGBonds  loops through the .pdb files in the directory; however, this process neds to be revised since the [glob *.pdb] list generates a wrong order. The current alternative is to read an input file generated in the Linux Shell (since globbing does seem to order the files properly).
GraphStrands_movie : A minimal update adding fixing a comment.